refactor playlists list
