+++ KaraIOke/ViewModels/AddingViewModel.cs
@@ -1,0 +1,12 @@
+namespace KaraIOke.ViewModels;
+
+public class AddingViewModel : AbstractPlaylistViewModel
+{
+    //TODO : Command to add a song to the playlist
+
+
+    public AddingViewModel(IServiceProvider serviceProvider) : base(serviceProvider)
+    {
+        //TODO : Initialize the command to add a song to the playlist
+    }
+}
